add file storage, add missing interface functions
